@@ -1,12 +1,14 @@
 import 'package:flutter/material.dart';
 
 class ChatMessage extends StatelessWidget {
-
-  // ignore: avoid_positional_boolean_parameters
-  ChatMessage(this.data, this.mine);
+  const ChatMessage({
+    super.key,
+    required this.data,
+    required this.isMine,
+  });
 
   final Map<String, dynamic> data;
-  final bool mine;
+  final bool isMine;
 
   @override
   Widget build(BuildContext context) {
@@ -14,45 +16,51 @@ class ChatMessage extends StatelessWidget {
       margin: const EdgeInsets.symmetric(vertical: 8, horizontal: 10),
       child: Row(
         children: <Widget>[
-          !mine ?
-          Padding(
-            padding: const EdgeInsets.only(right: 16),
-            child: CircleAvatar(
-              backgroundImage: NetworkImage(data['senderPhotoUrl']),
-            ),
-          ) : Container(),
-          Expanded(
-            child: Column(
-              crossAxisAlignment: mine 
-                ? CrossAxisAlignment.end 
-                : CrossAxisAlignment.start,
-              children: <Widget>[
-                data['imgUrl'] != null 
-                  ? Image.network(data['imgUrl'], width: 250)
-                  : Text(
-                    data['text'],
-                    textAlign: mine 
-                    ? TextAlign.end 
-                    : TextAlign.start,
-                    style: TextStyle(fontSize: 18),
-                  ),
-                  Text(
-                    data['senderName'],
-                    style: TextStyle(
-                      fontSize: 15,
-                      fontWeight: FontWeight.w500,
+          isMine
+              ? const SizedBox()
+              : Padding(
+                  padding: const EdgeInsets.only(right: 16),
+                  child: CircleAvatar(
+                    backgroundImage: NetworkImage(
+                      data['senderPhotoUrl'],
                     ),
                   ),
+                ),
+          Expanded(
+            child: Column(
+              crossAxisAlignment:
+                  isMine ? CrossAxisAlignment.end : CrossAxisAlignment.start,
+              children: <Widget>[
+                data['imgUrl'] != null
+                    ? Image.network(
+                        data['imgUrl'],
+                        width: 250,
+                      )
+                    : Text(
+                        data['text'],
+                        textAlign: isMine ? TextAlign.end : TextAlign.start,
+                        style: const TextStyle(fontSize: 18),
+                      ),
+                Text(
+                  data['senderName'],
+                  style: const TextStyle(
+                    fontSize: 15,
+                    fontWeight: FontWeight.w500,
+                  ),
+                ),
               ],
             ),
           ),
-          mine ?
-          Padding(
-            padding: const EdgeInsets.only(left: 16),
-            child: CircleAvatar(
-              backgroundImage: NetworkImage(data['senderPhotoUrl']),
-            ),
-          ) : Container(),
+          isMine
+              ? Padding(
+                  padding: const EdgeInsets.only(left: 16),
+                  child: CircleAvatar(
+                    backgroundImage: NetworkImage(
+                      data['senderPhotoUrl'],
+                    ),
+                  ),
+                )
+              : const SizedBox(),
         ],
       ),
     );
