@@ -1,21 +1,10 @@
 import 'package:flutter/material.dart';
-import 'chat_screen.dart';
 
-void main() async {
-  runApp(MyApp());
-}
+import 'chat_app.dart';
 
-class MyApp extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp(
-      title: 'Chat Flutter',
-      debugShowCheckedModeBanner: false,
-      home: ChatScreen(),
-      theme: ThemeData(
-        primarySwatch: Colors.green,
-        iconTheme: IconThemeData(color: Colors.green),
-      ),
-    );
-  }
+Future<void> main() async {
+  WidgetsFlutterBinding.ensureInitialized();
+  runApp(
+    const ChatApp(),
+  );
 }
