@@ -1,21 +1,26 @@
 import 'dart:io';
 
 import 'package:flutter/material.dart';
+
 import 'package:image_picker/image_picker.dart';
 
 class TextComposer extends StatefulWidget {
-  TextComposer(this.sendMessage);
-  final Function({String text, File imgFile}) sendMessage;
+  const TextComposer({
+    super.key,
+    required this.sendMessage,
+  });
+
+  final Function({String? text, File? imgFile}) sendMessage;
 
   @override
-  _TextComposerState createState() => _TextComposerState();
+  State<StatefulWidget> createState() => _TextComposerState();
 }
 
 class _TextComposerState extends State<TextComposer> {
-  final TextEditingController _controller = TextEditingController();
+  final _controller = TextEditingController();
   bool _isComposing = false;
 
-  void _reset(){
+  void _reset() {
     _controller.clear();
     setState(() {
       _isComposing = false;
@@ -29,37 +34,41 @@ class _TextComposerState extends State<TextComposer> {
       child: Row(
         children: <Widget>[
           IconButton(
-            icon: Icon(Icons.photo_camera),
+            icon: const Icon(Icons.photo_camera),
             onPressed: () async {
-              final imgFile =
-              // ignore: deprecated_member_use
-              await ImagePicker.pickImage(source: ImageSource.gallery);
+              final imgFile = await ImagePicker().pickImage(
+                source: ImageSource.gallery,
+              );
+              final selectedImage = File(imgFile?.path ?? '');
               if (imgFile == null) return;
-              widget.sendMessage(imgFile: imgFile);
+              widget.sendMessage(imgFile: selectedImage);
             },
           ),
           Expanded(
             child: TextField(
               controller: _controller,
-              decoration: InputDecoration.collapsed(
-                hintText: 'Enviar mensagem...'),
-              onChanged: (text){
+              decoration: const InputDecoration.collapsed(
+                hintText: 'Enviar mensagem...',
+              ),
+              onChanged: (text) {
                 setState(() {
                   _isComposing = text.isNotEmpty;
                 });
               },
-              onSubmitted: (text){
+              onSubmitted: (text) {
                 widget.sendMessage(text: text);
                 _reset();
               },
             ),
           ),
           IconButton(
-            icon: Icon(Icons.send, color: Colors.green),
-            onPressed: _isComposing ? (){
-              widget.sendMessage(text: _controller.text);
-              _reset();
-            } : null,
+            icon: const Icon(Icons.send, color: Colors.green),
+            onPressed: _isComposing
+                ? () {
+                    widget.sendMessage(text: _controller.text);
+                    _reset();
+                  }
+                : null,
           ),
         ],
       ),
